checker/main_test: init
This adds unit tests for checker.Run.

TestChecker tests Healthcheck.Run.
It passes in a mock.Healthcheck and a dummySrvManager, then
uses the mock.Healthcheck.HealthC channel to change health
and inspects dummySrvManagers rrset afterwards.
It also tests the record is removed from the set once the context is
cancelled.