main.go: Instantiate Dummy SRV manager when --dry-run is passed
Also, remove the "dry run" mode in the Route53 SRV Manager.

There's few benefit in a dry-run that verifies it can read (but not
write) DNS records in Route53, and making the --dry-run flag actually
use the dummy srv manager (which does log) allows smoketesting it
without needing to be able to connect to Route53.